LPS-84329 Extracting notifications code to its own module
