+++ modules/apps/sharing/sharing-notifications/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Sharing Notifications
+Bundle-SymbolicName: com.liferay.sharing.notifications
+Bundle-Version: 1.0.0
+Web-ContextPath: /sharing-notifications
