+++ modules/apps/sharing/sharing-notifications/src/main/java/com/liferay/sharing/notifications/internal/notifications/AddSharingEntryUserNotificationDefinition.java
@@ -12,12 +12,12 @@
  * details.
  */
 
-package com.liferay.sharing.web.internal.notifications;
+package com.liferay.sharing.notifications.internal.notifications;
 
 import com.liferay.portal.kernel.model.UserNotificationDeliveryConstants;
 import com.liferay.portal.kernel.notifications.UserNotificationDefinition;
 import com.liferay.portal.kernel.notifications.UserNotificationDeliveryType;
-import com.liferay.sharing.web.internal.constants.SharingPortletKeys;
+import com.liferay.sharing.constants.SharingPortletKeys;
 
 import org.osgi.service.component.annotations.Component;
