+++ modules/apps/sharing/sharing-notifications/src/main/resources/content/Language.properties
@@ -1,0 +1,7 @@
+commenting=commenting
+editing=editing
+receive-a-notification-when-someone-shares-content-with-you=Receives a notification when someone shares content with you.
+someone=Someone
+something=something
+viewing=viewing
+x-has-shared-x-with-you-for-x={0} has shared {1} with you for {2}.
