+++ modules/apps/sharing/sharing-web/src/main/java/com/liferay/sharing/web/internal/portlet/action/ShareEntryMVCActionCommand.java
@@ -25,8 +25,8 @@ import com.liferay.portal.kernel.util.DateFormatFactoryUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.sharing.service.SharingEntryService;
 import com.liferay.sharing.constants.SharingPortletKeys;
+import com.liferay.sharing.service.SharingEntryService;
 import com.liferay.sharing.web.internal.display.SharingEntryPermissionDisplayActionKey;
 
 import java.util.Date;
