+++ modules/apps/sharing/sharing-web/src/main/resources/content/Language.properties
@@ -2,15 +2,8 @@ allow-the-document-to-be-shared-with-other-users=Allow the document to be shared
 collaborators-can-comment-on-the-document=Collaborators can comment on the document.
 collaborators-can-only-view-the-document=Collaborators can only view the document.
 collaborators-can-view-comment-update-the-document=Collaborators can view, comment, and update the document.
-commenting=commenting
-editing=editing
 edition=Edition
 invite-to-collaborate=Invite to Collaborate
 javax.portlet.title.com_liferay_sharing_web_portlet_SharingPortlet=Sharing
-receive-a-notification-when-someone-shares-content-with-you=Receives a notification when someone shares content with you.
 sharing-permissions=Sharing Permissions
-someone=Someone
-something=something
 use-commas-to-include-multiple-collaborators=Use commas to include multiple collaborators.
-viewing=viewing
-x-has-shared-x-with-you-for-x={0} has shared {1} with you for {2}.
